fix(cycle): retire fix strategies per-finding, not per-cycle
runFixCycle retired a strategy for the whole cycle when it answered UNRESOLVED
(#1369). But UNRESOLVED is a per-finding signal: the agent said "I cannot fix
THIS", not "I cannot fix anything in this story". When a later, unrelated finding
routed only to that retired strategy, the cycle exited `no-strategy` and the
caller discarded the whole pass — even when the outstanding fix was trivial.

On otel-telemetry-expansion US-006 that cost ~9 minutes and 3 agent sessions: the
implementer correctly refused finding A (it needed a scoping decision it could not
guess), and the one-line barrel export its own co-run sibling had just created was
then orphaned, rolling back the sibling's commit. `createBatchQueue` shipped to
main as dead code as a result.

Retirement is now scoped to (strategy, finding) via a decline ledger. The declined
unit is the DISPATCHED BATCH — a strategy receives every finding it claims in one
call and answers once — which over-retires within the batch, harmlessly and in the
conservative direction. Termination is preserved: a strategy declines any finding
at most once, the declined set only grows, and the attempt caps are untouched.

`findingKey` includes `message`, so a re-emitted finding with rephrased text mints
a new key and the strategy becomes dispatchable again. Accepted rather than
worked around: the drift is in the "try again" direction, mechanical producers have
stable messages, and maxAttemptsPerStrategy still binds. Pinned by a test.

The all-gave-up early exit deliberately stays: nothing in the group touched the
tree, so validate can only re-emit findings already in the declined batch — there
is no finding to re-dispatch FOR. It reads as a contradiction of this change, so
the comment now says why it is not.

`retiredStrategies` in the orphan-exit log now means "retired with respect to the
remaining findings", which is the diagnostic that exit actually needs.

Splits: the ledger moves to src/findings/cycle-retirement.ts and the retirement
tests to test/unit/findings/cycle-retirement.test.ts, both of which had breached
the file-size caps (600 src / 800 test). Shared cycle fixtures move to
_cycle-fixtures.ts so the two test files cannot drift apart.

Refs #1384
